Add files via upload
This system helps the system to keep the book records. in this project, you can add books and can issue them to the registered users. Also, the admin can add different books and can delete them later. The admin has to first register the student before issuing the book. You can view the book details that are available in the library.
The library management system is a simple and totally complete book library system. This system requires various kinds of jar files in order to run it smoothly. And the so MySQL database as backend. This system whole is designed in NetBeans IDE. To run this project you can use both NetBeans IDE or Eclipse IDE.